+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-evaluator-impl/src/main/java/com/liferay/dynamic/data/mapping/form/evaluator/impl/internal/DDMFormEvaluatorHelper.java
@@ -20,9 +20,12 @@ import com.liferay.dynamic.data.mapping.expression.DDMExpressionException;
 import com.liferay.dynamic.data.mapping.expression.DDMExpressionFactory;
 import com.liferay.dynamic.data.mapping.form.evaluator.DDMFormEvaluationException;
 import com.liferay.dynamic.data.mapping.form.evaluator.DDMFormEvaluationResult;
+import com.liferay.dynamic.data.mapping.form.evaluator.DDMFormEvaluatorContext;
 import com.liferay.dynamic.data.mapping.form.evaluator.DDMFormFieldEvaluationResult;
+import com.liferay.dynamic.data.mapping.form.evaluator.impl.internal.functions.BelongsToRoleFunction;
 import com.liferay.dynamic.data.mapping.form.evaluator.impl.internal.functions.CallFunction;
 import com.liferay.dynamic.data.mapping.form.evaluator.impl.internal.functions.GetPropertyFunction;
+import com.liferay.dynamic.data.mapping.form.evaluator.impl.internal.functions.JumpPageFunction;
 import com.liferay.dynamic.data.mapping.form.evaluator.impl.internal.functions.SetEnabledFunction;
 import com.liferay.dynamic.data.mapping.form.evaluator.impl.internal.functions.SetInvalidFunction;
 import com.liferay.dynamic.data.mapping.form.evaluator.impl.internal.functions.SetPropertyFunction;
@@ -42,6 +45,7 @@ import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -49,11 +53,15 @@ import com.liferay.portal.kernel.util.Validator;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
+
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * @author Leonardo Barros
@@ -63,21 +71,26 @@ public class DDMFormEvaluatorHelper {
 	public DDMFormEvaluatorHelper(
 		DDMDataProviderTracker ddmDataProviderTracker,
 		DDMDataProviderInstanceService ddmDataProviderInstanceService,
-		DDMExpressionFactory ddmExpressionFactory, DDMForm ddmForm,
-		DDMFormValues ddmFormValues,
+		DDMExpressionFactory ddmExpressionFactory,
+		DDMFormEvaluatorContext ddmFormEvaluatorContext,
 		DDMFormValuesJSONDeserializer ddmFormValuesJSONDeserializer,
-		JSONFactory jsonFactory, Locale locale) {
+		JSONFactory jsonFactory, UserLocalService userLocalService) {
 
 		_ddmDataProviderTracker = ddmDataProviderTracker;
 		_ddmDataProviderInstanceService = ddmDataProviderInstanceService;
 		_ddmExpressionFactory = ddmExpressionFactory;
-		_ddmForm = ddmForm;
-		_ddmFormFieldsMap = ddmForm.getDDMFormFieldsMap(true);
+		_ddmForm = ddmFormEvaluatorContext.getDDMForm();
+
+		_ddmFormFieldsMap = _ddmForm.getDDMFormFieldsMap(true);
+
 		_ddmFormValuesJSONDeserializer = ddmFormValuesJSONDeserializer;
 		_jsonFactory = jsonFactory;
-		_locale = locale;
+		_userLocalService = userLocalService;
+		_locale = ddmFormEvaluatorContext.getLocale();
 
-		createDDMFormFieldValues(ddmFormValues);
+		_request = ddmFormEvaluatorContext.getProperty("request");
+
+		createDDMFormFieldValues(ddmFormEvaluatorContext.getDDMFormValues());
 
 		createDDMFormFieldRuleEvaluationResultsMap();
 	}
@@ -100,6 +113,10 @@ public class DDMFormEvaluatorHelper {
 
 		ddmFormEvaluationResult.setDDMFormFieldEvaluationResults(
 			ddmFormFieldEvaluationResults);
+
+		Set<Integer> disabledPagesIndexes = getDisabledPagesIndexes();
+
+		ddmFormEvaluationResult.setDisabledPagesIndexes(disabledPagesIndexes);
 
 		return ddmFormEvaluationResult;
 	}
@@ -158,6 +175,20 @@ public class DDMFormEvaluatorHelper {
 
 			ddmFormFieldEvaluationResultInstances.add(
 				ddmFormFieldEvaluationResult);
+
+			for (DDMFormFieldValue nestedDDMFormFieldValue :
+					ddmFormFieldValue.getNestedDDMFormFieldValues()) {
+
+				DDMFormField nestedDDMFormField =
+					nestedDDMFormFieldValue.getDDMFormField();
+
+				ddmFormFieldEvaluationResult =
+					createDDMFormFieldEvaluationResult(
+						nestedDDMFormField, nestedDDMFormFieldValue);
+
+				ddmFormFieldEvaluationResultInstances.add(
+					ddmFormFieldEvaluationResult);
+			}
 		}
 
 		_ddmFormFieldEvaluationResultsMap.put(
@@ -215,6 +246,16 @@ public class DDMFormEvaluatorHelper {
 			if (instanceId.equals(ddmFormFieldValue.getInstanceId())) {
 				return ddmFormFieldValue;
 			}
+
+			for (DDMFormFieldValue nestedDDMFormFieldValue :
+					ddmFormFieldValue.getNestedDDMFormFieldValues()) {
+
+				if (instanceId.equals(
+						nestedDDMFormFieldValue.getInstanceId())) {
+
+					return nestedDDMFormFieldValue;
+				}
+			}
 		}
 
 		return null;
@@ -235,6 +276,21 @@ public class DDMFormEvaluatorHelper {
 		}
 
 		return defaultValue;
+	}
+
+	protected Set<Integer> getDisabledPagesIndexes() {
+		Set<Integer> disabledPagesIndexes = new HashSet<>();
+
+		for (Map.Entry<Integer, Integer> entry : _pageFlow.entrySet()) {
+			int fromPageIndex = entry.getKey();
+			int toPageIndex = entry.getValue();
+
+			for (int i = fromPageIndex + 1; i < toPageIndex; i++) {
+				disabledPagesIndexes.add(i);
+			}
+		}
+
+		return disabledPagesIndexes;
 	}
 
 	protected String getJSONArrayValueString(String valueString) {
@@ -321,6 +377,9 @@ public class DDMFormEvaluatorHelper {
 		DDMFormRuleEvaluator ddmFormRuleEvaluator) {
 
 		ddmFormRuleEvaluator.setDDMExpressionFunction(
+			"belongsTo",
+			new BelongsToRoleFunction(_request, _userLocalService));
+		ddmFormRuleEvaluator.setDDMExpressionFunction(
 			"call",
 			new CallFunction(
 				_ddmDataProviderTracker, _ddmDataProviderInstanceService,
@@ -330,6 +389,8 @@ public class DDMFormEvaluatorHelper {
 			"getValue",
 			new GetPropertyFunction(
 				_ddmFormFieldEvaluationResultsMap, "value"));
+		ddmFormRuleEvaluator.setDDMExpressionFunction(
+			"jumpPage", new JumpPageFunction(_pageFlow));
 		ddmFormRuleEvaluator.setDDMExpressionFunction(
 			"setEnabled",
 			new SetEnabledFunction(_ddmFormFieldEvaluationResultsMap));
@@ -546,5 +607,8 @@ public class DDMFormEvaluatorHelper {
 	private final DDMFormValuesJSONDeserializer _ddmFormValuesJSONDeserializer;
 	private final JSONFactory _jsonFactory;
 	private final Locale _locale;
+	private final Map<Integer, Integer> _pageFlow = new HashMap<>();
+	private final HttpServletRequest _request;
+	private final UserLocalService _userLocalService;
 
 }