+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/service/impl/DDMTemplateLocalServiceImpl.java
@@ -495,7 +495,7 @@ public class DDMTemplateLocalServiceImpl
 	public DDMTemplate fetchTemplate(
 		long groupId, long classNameId, String templateKey) {
 
-		templateKey = StringUtil.toUpperCase(templateKey.trim());
+		templateKey = StringUtil.toUpperCase(StringUtil.trim(templateKey));
 
 		return ddmTemplatePersistence.fetchByG_C_T(
 			groupId, classNameId, templateKey);
@@ -530,7 +530,7 @@ public class DDMTemplateLocalServiceImpl
 			boolean includeAncestorTemplates)
 		throws PortalException {
 
-		templateKey = StringUtil.toUpperCase(templateKey.trim());
+		templateKey = StringUtil.toUpperCase(StringUtil.trim(templateKey));
 
 		DDMTemplate template = ddmTemplatePersistence.fetchByG_C_T(
 			groupId, classNameId, templateKey);
@@ -584,7 +584,7 @@ public class DDMTemplateLocalServiceImpl
 			long groupId, long classNameId, String templateKey)
 		throws PortalException {
 
-		templateKey = StringUtil.toUpperCase(templateKey.trim());
+		templateKey = StringUtil.toUpperCase(StringUtil.trim(templateKey));
 
 		return ddmTemplatePersistence.findByG_C_T(
 			groupId, classNameId, templateKey);
@@ -618,7 +618,7 @@ public class DDMTemplateLocalServiceImpl
 			boolean includeAncestorTemplates)
 		throws PortalException {
 
-		templateKey = StringUtil.toUpperCase(templateKey.trim());
+		templateKey = StringUtil.toUpperCase(StringUtil.trim(templateKey));
 
 		DDMTemplate template = ddmTemplatePersistence.fetchByG_C_T(
 			groupId, classNameId, templateKey);
@@ -1335,11 +1335,11 @@ public class DDMTemplateLocalServiceImpl
 	 */
 	@Override
 	public DDMTemplate updateTemplate(
-			long userId, long templateId, long classPK, Map<Locale,
-			String> nameMap, Map<Locale, String> descriptionMap, String type,
-			String mode, String language, String script, boolean cacheable,
-			boolean smallImage, String smallImageURL, File smallImageFile,
-			ServiceContext serviceContext)
+			long userId, long templateId, long classPK,
+			Map<Locale, String> nameMap, Map<Locale, String> descriptionMap,
+			String type, String mode, String language, String script,
+			boolean cacheable, boolean smallImage, String smallImageURL,
+			File smallImageFile, ServiceContext serviceContext)
 		throws PortalException {
 
 		User user = userLocalService.getUser(userId);
@@ -1594,7 +1594,7 @@ public class DDMTemplateLocalServiceImpl
 			String smallImageURL, File smallImageFile, byte[] smallImageBytes)
 		throws PortalException {
 
-		templateKey = StringUtil.toUpperCase(templateKey.trim());
+		templateKey = StringUtil.toUpperCase(StringUtil.trim(templateKey));
 
 		DDMTemplate template = ddmTemplatePersistence.fetchByG_C_T(
 			groupId, classNameId, templateKey);
@@ -1634,8 +1634,7 @@ public class DDMTemplateLocalServiceImpl
 				ddmGroupServiceConfiguration.smallImageExtensions()) {
 
 			if (StringPool.STAR.equals(smallImageExtension) ||
-				StringUtil.endsWith(
-					smallImageName, smallImageExtension)) {
+				StringUtil.endsWith(smallImageName, smallImageExtension)) {
 
 				validSmallImageExtension = true;